tree: introduce header file for output
The file contains only one macro right now. But it's more convenient
if the header with the appropriate name will do output routines.

This is one of the steps on the way to get rid of errprint_*() macros
and handle errors by yourself. For more context see other patches
under the same topic specified below.

TOPIC=drop_errprint